remove fetching logic from leaderboard
