+++ osu.Game/Screens/Select/PlayBeatmapDetailArea.cs
@@ -1,8 +1,6 @@
 // Copyright (c) ppy Pty Ltd <contact@ppy.sh>. Licensed under the MIT Licence.
 // See the LICENCE file in the repository root for full licence text.
 
-#nullable disable
-
 using System;
 using System.Linq;
 using osu.Framework.Allocation;
@@ -10,13 +8,18 @@ using osu.Framework.Bindables;
 using osu.Framework.Graphics;
 using osu.Game.Beatmaps;
 using osu.Game.Configuration;
+using osu.Game.Online.Leaderboards;
+using osu.Game.Scoring;
 using osu.Game.Screens.Select.Leaderboards;
 
 namespace osu.Game.Screens.Select
 {
     public partial class PlayBeatmapDetailArea : BeatmapDetailArea
     {
-        public readonly BeatmapLeaderboard Leaderboard;
+        [Resolved]
+        private LeaderboardScoresProvider<BeatmapLeaderboardScope, ScoreInfo> leaderboardProvider { get; set; } = null!;
+
+        private BeatmapLeaderboard leaderboard;
 
         public override WorkingBeatmap Beatmap
         {
@@ -25,17 +28,21 @@ namespace osu.Game.Screens.Select
             {
                 base.Beatmap = value;
 
-                Leaderboard.BeatmapInfo = value is DummyWorkingBeatmap ? null : value?.BeatmapInfo;
+                ((BeatmapLeaderboardScoresProvider)leaderboardProvider).BeatmapInfo = value is DummyWorkingBeatmap ? null : value?.BeatmapInfo;
             }
         }
 
-        private Bindable<TabType> selectedTab;
+        private Bindable<TabType> selectedTab = null!;
 
-        private Bindable<bool> selectedModsFilter;
+        private Bindable<bool> selectedModsFilter = null!;
 
-        public PlayBeatmapDetailArea()
+        public PlayBeatmapDetailArea(Action<ScoreInfo>? onScoreSelected = null)
         {
-            Add(Leaderboard = new BeatmapLeaderboard { RelativeSizeAxes = Axes.Both });
+            Add(leaderboard = new BeatmapLeaderboard
+            {
+                ScoreSelected = onScoreSelected,
+                RelativeSizeAxes = Axes.Both
+            });
         }
 
         [BackgroundDependencyLoader]
@@ -55,24 +62,24 @@ namespace osu.Game.Screens.Select
         {
             base.Refresh();
 
-            Leaderboard.RefetchScores();
+            leaderboardProvider.RefetchScores();
         }
 
         protected override void OnTabChanged(BeatmapDetailAreaTabItem tab, bool selectedMods)
         {
             base.OnTabChanged(tab, selectedMods);
 
-            Leaderboard.FilterMods = selectedMods;
+            ((BeatmapLeaderboardScoresProvider)leaderboardProvider).FilterMods = selectedMods;
 
             switch (tab)
             {
                 case BeatmapDetailAreaLeaderboardTabItem<BeatmapLeaderboardScope> leaderboard:
-                    Leaderboard.Scope = leaderboard.Scope;
-                    Leaderboard.Show();
+                    leaderboardProvider.Scope = leaderboard.Scope;
+                    this.leaderboard.Show();
                     break;
 
                 default:
-                    Leaderboard.Hide();
+                    leaderboard.Hide();
                     break;
             }
         }
